requirements.txt
Click==7.0
dj-database-url==0.5.0
Django==2.2.2
django-heroku==0.3.1
djangorestframework==3.9.4
Flask==1.0.3
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2==2.8.2
Pygments==2.4.2
pytz==2019.1
sqlparse==0.3.0
Werkzeug==0.15.4
whitenoise==4.1.2
